フッターに世界観を統一したデザインを追加

+++ app/controllers/relationships_controller.rb
@@ -2,14 +2,25 @@ class RelationshipsController < ApplicationController
   before_action :authenticate_user!
 
   def create
-    user = User.find(params[:followed_id])
-    current_user.follow(user)
-    redirect_back(fallback_location: root_path)
+    @user = User.find(params[:followed_id])
+    current_user.follow(@user)
+
+    respond_to do |format|
+      format.js
+    end
   end
 
   def destroy
-    relationship = current_user.active_relationships.find(params[:id])
+    relationship = Relationship.find(params[:id])
+
+    # 念のためのガード
+    return head :forbidden unless relationship.follower == current_user
+
+    @user = relationship.followed
     relationship.destroy
-    redirect_back(fallback_location: root_path)
+
+    respond_to do |format|
+      format.js
+    end
   end
 end
